requirements.txt
bidict==0.21.2
cachelib==0.1.1
click==7.1.2
Flask==1.1.2
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
gunicorn==20.0.4
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
python-engineio==4.0.0
python-socketio==5.0.3
scipy==1.5.4
Shapely==1.7.1
waitress==1.4.4
Werkzeug==1.0.1
